refactor(commands): extract whoami tests to sibling file
